refactor(structure): reorganize files into respective folders and update import paths
- Moved files to appropriate directories based on functionality
- Updated import paths in related files to reflect the new structure
- Improved project organization for better maintainability
