+++ src/auth/auth.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
-import { AuthController } from './auth.controller';
-import { AuthService } from './auth.service';
+import { AuthController } from './controller/auth.controller';
+import { AuthService } from './service/auth.service';
 import { PrismaConnectionModule } from '../prisma-connection/prisma-connection.module';
 import { JwtModule } from '@nestjs/jwt';
 import { JwtStrategy } from '../strategy';
